Revert "logs/pattern: add reusable signature + delegate internal tokenizer (moved from automultiline)"
This reverts commit eea3ea1c73c603b04050fd20d2f5701e8b97c651.
