+++ comp/observer/impl/log_timeseries_analysis.go
@@ -14,7 +14,6 @@ import (
 	"unicode"
 
 	observer "github.com/DataDog/datadog-agent/comp/observer/def"
-	"github.com/DataDog/datadog-agent/pkg/logs/pattern"
 )
 
 // LogTimeSeriesAnalysis converts logs into timeseries metric outputs:
@@ -44,7 +43,7 @@ func (a *LogTimeSeriesAnalysis) Analyze(log observer.LogView) observer.LogAnalys
 	}
 
 	// Unstructured path: pattern frequency
-	patternSig := pattern.Signature(content, a.MaxEvalBytes)
+	patternSig := logSignature(content, a.MaxEvalBytes)
 	if patternSig == "" {
 		return observer.LogAnalysisResult{}
 	}
@@ -78,7 +77,7 @@ func (a *LogTimeSeriesAnalysis) extractJSONMetrics(content []byte, tags []string
 	// Pattern frequency for structured logs:
 	// Always derive a signature from the rendered JSON payload so ALL structured logs emit a metric,
 	// regardless of whether numeric fields are present.
-	if sig := pattern.Signature(content, a.MaxEvalBytes); sig != "" {
+	if sig := logSignature(content, a.MaxEvalBytes); sig != "" {
 		out = append(out, observer.MetricOutput{
 			Name:        patternCountMetricName(sig),
 			Value:       1,
@@ -170,4 +169,332 @@ func sanitizeMetricFragment(s string) string {
 		}
 	}
 	return b.String()
+}
+
+// ---------------------------------------------------------------------------
+// Inlined log signature generation (avoids dependency on pkg/logs/pattern)
+// ---------------------------------------------------------------------------
+
+// maxRun is the maximum run of a char or digit before it is capped.
+const maxRun = 10
+
+// Token types for log signature generation
+type sigToken byte
+
+const (
+	tokSpace sigToken = iota
+	tokColon
+	tokSemicolon
+	tokDash
+	tokUnderscore
+	tokFslash
+	tokBslash
+	tokPeriod
+	tokComma
+	tokSinglequote
+	tokDoublequote
+	tokBacktick
+	tokTilda
+	tokStar
+	tokPlus
+	tokEqual
+	tokParenopen
+	tokParenclose
+	tokBraceopen
+	tokBraceclose
+	tokBracketopen
+	tokBracketclose
+	tokAmpersand
+	tokExclamation
+	tokAt
+	tokPound
+	tokDollar
+	tokPercent
+	tokUparrow
+	// Digit runs D1..D10
+	tokD1
+	tokD2
+	tokD3
+	tokD4
+	tokD5
+	tokD6
+	tokD7
+	tokD8
+	tokD9
+	tokD10
+	// Char runs C1..C10
+	tokC1
+	tokC2
+	tokC3
+	tokC4
+	tokC5
+	tokC6
+	tokC7
+	tokC8
+	tokC9
+	tokC10
+	// Special tokens
+	tokMonth
+	tokDay
+	tokApm
+	tokZone
+	tokT
+	tokEnd
+)
+
+// logSignature returns a deterministic signature for the input bytes, capped to maxEvalBytes if > 0.
+func logSignature(input []byte, maxEvalBytes int) string {
+	if len(input) == 0 {
+		return ""
+	}
+	maxBytes := len(input)
+	if maxEvalBytes > 0 && maxBytes > maxEvalBytes {
+		maxBytes = maxEvalBytes
+	}
+	input = input[:maxBytes]
+
+	var out strings.Builder
+	out.Grow(len(input))
+
+	var buf [maxRun]byte
+	bufLen := 0
+	resetBuf := func() { bufLen = 0 }
+	appendBuf := func(b byte) {
+		if bufLen < len(buf) {
+			if b >= 'a' && b <= 'z' {
+				b = b - 'a' + 'A'
+			}
+			buf[bufLen] = b
+			bufLen++
+		}
+	}
+
+	run := 0
+	lastToken := getTokenType(input[0])
+	resetBuf()
+	appendBuf(input[0])
+
+	insertToken := func() {
+		defer func() {
+			run = 0
+			resetBuf()
+		}()
+
+		if lastToken == tokC1 {
+			if bufLen == 1 {
+				if specialToken := getSpecialShortToken(buf[0]); specialToken != tokEnd {
+					out.WriteString(sigTokenToString(specialToken))
+					return
+				}
+			} else if bufLen > 1 {
+				if specialToken := getSpecialLongToken(string(buf[:bufLen])); specialToken != tokEnd {
+					out.WriteString(sigTokenToString(specialToken))
+					return
+				}
+			}
+		}
+
+		if lastToken == tokC1 || lastToken == tokD1 {
+			if run >= maxRun {
+				run = maxRun - 1
+			}
+			out.WriteString(sigTokenToString(lastToken + sigToken(run)))
+			return
+		}
+
+		out.WriteString(sigTokenToString(lastToken))
+	}
+
+	for _, char := range input[1:] {
+		currentToken := getTokenType(char)
+		if currentToken != lastToken {
+			insertToken()
+		} else {
+			run++
+		}
+		appendBuf(char)
+		lastToken = currentToken
+	}
+	insertToken()
+
+	return out.String()
+}
+
+func getTokenType(char byte) sigToken {
+	if unicode.IsDigit(rune(char)) {
+		return tokD1
+	} else if unicode.IsSpace(rune(char)) {
+		return tokSpace
+	}
+
+	switch char {
+	case ':':
+		return tokColon
+	case ';':
+		return tokSemicolon
+	case '-':
+		return tokDash
+	case '_':
+		return tokUnderscore
+	case '/':
+		return tokFslash
+	case '\\':
+		return tokBslash
+	case '.':
+		return tokPeriod
+	case ',':
+		return tokComma
+	case '\'':
+		return tokSinglequote
+	case '"':
+		return tokDoublequote
+	case '`':
+		return tokBacktick
+	case '~':
+		return tokTilda
+	case '*':
+		return tokStar
+	case '+':
+		return tokPlus
+	case '=':
+		return tokEqual
+	case '(':
+		return tokParenopen
+	case ')':
+		return tokParenclose
+	case '{':
+		return tokBraceopen
+	case '}':
+		return tokBraceclose
+	case '[':
+		return tokBracketopen
+	case ']':
+		return tokBracketclose
+	case '&':
+		return tokAmpersand
+	case '!':
+		return tokExclamation
+	case '@':
+		return tokAt
+	case '#':
+		return tokPound
+	case '$':
+		return tokDollar
+	case '%':
+		return tokPercent
+	case '^':
+		return tokUparrow
+	}
+
+	return tokC1
+}
+
+func getSpecialShortToken(char byte) sigToken {
+	switch char {
+	case 'T':
+		return tokT
+	case 'Z':
+		return tokZone
+	}
+	return tokEnd
+}
+
+func getSpecialLongToken(input string) sigToken {
+	switch input {
+	case "JAN", "FEB", "MAR", "APR", "MAY", "JUN", "JUL",
+		"AUG", "SEP", "OCT", "NOV", "DEC":
+		return tokMonth
+	case "MON", "TUE", "WED", "THU", "FRI", "SAT", "SUN":
+		return tokDay
+	case "AM", "PM":
+		return tokApm
+	case "UTC", "GMT", "EST", "EDT", "CST", "CDT",
+		"MST", "MDT", "PST", "PDT", "JST", "KST",
+		"IST", "MSK", "CEST", "CET", "BST", "NZST",
+		"NZDT", "ACST", "ACDT", "AEST", "AEDT",
+		"AWST", "AWDT", "AKST", "AKDT", "HST",
+		"HDT", "CHST", "CHDT", "NST", "NDT":
+		return tokZone
+	}
+	return tokEnd
+}
+
+func sigTokenToString(token sigToken) string {
+	if token >= tokD1 && token <= tokD10 {
+		return strings.Repeat("D", int(token-tokD1)+1)
+	} else if token >= tokC1 && token <= tokC10 {
+		return strings.Repeat("C", int(token-tokC1)+1)
+	}
+
+	switch token {
+	case tokSpace:
+		return " "
+	case tokColon:
+		return ":"
+	case tokSemicolon:
+		return ";"
+	case tokDash:
+		return "-"
+	case tokUnderscore:
+		return "_"
+	case tokFslash:
+		return "/"
+	case tokBslash:
+		return "\\"
+	case tokPeriod:
+		return "."
+	case tokComma:
+		return ","
+	case tokSinglequote:
+		return "'"
+	case tokDoublequote:
+		return "\""
+	case tokBacktick:
+		return "`"
+	case tokTilda:
+		return "~"
+	case tokStar:
+		return "*"
+	case tokPlus:
+		return "+"
+	case tokEqual:
+		return "="
+	case tokParenopen:
+		return "("
+	case tokParenclose:
+		return ")"
+	case tokBraceopen:
+		return "{"
+	case tokBraceclose:
+		return "}"
+	case tokBracketopen:
+		return "["
+	case tokBracketclose:
+		return "]"
+	case tokAmpersand:
+		return "&"
+	case tokExclamation:
+		return "!"
+	case tokAt:
+		return "@"
+	case tokPound:
+		return "#"
+	case tokDollar:
+		return "$"
+	case tokPercent:
+		return "%"
+	case tokUparrow:
+		return "^"
+	case tokMonth:
+		return "MTH"
+	case tokDay:
+		return "DAY"
+	case tokApm:
+		return "PM"
+	case tokT:
+		return "T"
+	case tokZone:
+		return "ZONE"
+	}
+	return ""
 }
